Use init generated files, clean up repo (#24)
Replaces the contents of the todo folder in this repo with the output of the init command from elastic/synthetics#525

This also adds examples using the request object filling a gap identified in the documentation of that feature (CC @colleenmcginnis )

It also removes the e-commerce demo temporarily till we have steady hosting for it. This blocks #25 as a result.

Finally, this adds inline examples, and removes the heartbeat folder, since as of 8.4 we want to discourage running browser monitors directly with heartbeat.